+++ plugins/crunch_compressor/converter/images_to_pdf_converter.py
@@ -27,7 +27,6 @@ possible_files_endings = ['.jpc', '.xbm', '.j2k', '.icns', '.png', '.bmp', '.jp2
 
 
 class ImagesToPdfConverter(Processor):
-
     def __init__(
             self,
             pytesseract_path: str = None,
